requirements.txt
openai==1.12.0
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.0
python-dotenv==1.0.1
flask==3.0.2